change memtable id from int to long

+++ server/src/main/java/org/apache/iotdb/db/engine/memtable/IMemTable.java
@@ -181,7 +181,7 @@ public interface IMemTable extends WALEntryValue {
 
   long getMinPlanIndex();
 
-  int getMemTableId();
+  long getMemTableId();
 
   long getCreatedTime();
